IOS-4765 Move PendingTransactionRecordMapper to manager
