+++ BlockchainSdk/Common/PendingTransactionRecord.swift
@@ -42,31 +42,3 @@ public struct PendingTransactionRecord {
         self.transactionParams = transactionParams
     }
 }
-
-struct PendingTransactionRecordMapper {
-    func makeDummy(blockchain: Blockchain) -> PendingTransactionRecord {
-        PendingTransactionRecord(
-            hash: .unknown,
-            source: .unknown,
-            destination: .unknown,
-            amount: .zeroCoin(for: blockchain),
-            fee: Fee(.zeroCoin(for: blockchain)),
-            date: Date(),
-            isIncoming: false,
-            transactionParams: nil
-        )
-    }
-    
-    func mapToPendingTransactionRecord(_ pendingTransaction: PendingTransaction, blockchain: Blockchain) -> PendingTransactionRecord {
-        PendingTransactionRecord(
-            hash: pendingTransaction.hash,
-            source: pendingTransaction.source,
-            destination: pendingTransaction.destination,
-            amount: Amount(with: blockchain, value: pendingTransaction.value),
-            fee: Fee(Amount(with: blockchain, value: pendingTransaction.fee ?? 0)),
-            date: pendingTransaction.date,
-            isIncoming: pendingTransaction.isIncoming,
-            transactionParams: pendingTransaction.transactionParams
-        )
-    }
-}
